Added the tests for 2 other DAOs. Added username and password generation features.

+++ src/main/java/ua/laboratory/lab_spring_task/Service/TrainingService.java
@@ -6,6 +6,6 @@ import java.util.List;
 
 public interface TrainingService {
     public Training createTraining(Training training);
-    public Training getTraining(int id);
+    public Training getTraining(Long id);
     public List<Training> getAllTrainings();
 }
